pyproject.toml
[project]
name = "bloomy-mcp"
version = "0.1.0"
description = "Bloom Growth GraphQL MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "pyyaml>=6.0.2",
    "gql>=3.4.1",
    "httpx>=0.27.0",
]

[project.scripts]
bloomy-server = "bloomy_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
